Remove Reactor dep, upgrade immutables, move encoding to own package
