+++ api/src/main/java/discord4j/discordjson/possible/PossibleDeserializer.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.deser.ValueInstantiator;
 import com.fasterxml.jackson.databind.deser.std.ReferenceTypeDeserializer;
 import com.fasterxml.jackson.databind.jsontype.TypeDeserializer;
-import reactor.util.annotation.Nullable;
+import javax.annotation.Nullable;
 
 import java.util.Optional;
