+++ api/src/main/java/discord4j/discordjson/possible/PossibleSerializer.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.jsontype.TypeSerializer;
 import com.fasterxml.jackson.databind.ser.std.ReferenceTypeSerializer;
 import com.fasterxml.jackson.databind.type.ReferenceType;
 import com.fasterxml.jackson.databind.util.NameTransformer;
-import reactor.util.annotation.Nullable;
+import javax.annotation.Nullable;
 
 public class PossibleSerializer extends ReferenceTypeSerializer<Possible<?>> {
