+++ encoding/src/main/java/discord4j/discordjson/encoding/possible/PossibleIdEncoding.java
@@ -1,6 +1,7 @@
-package discord4j.discordjson.possible;
+package discord4j.discordjson.encoding.possible;
 
 import discord4j.discordjson.Id;
+import discord4j.discordjson.possible.Possible;
 import org.immutables.encode.Encoding;
 
 import java.util.Objects;
