+++ encoding/src/main/java/discord4j/discordjson/encoding/possible/PossibleOptionalEncoding.java
@@ -1,7 +1,8 @@
-package discord4j.discordjson.possible;
+package discord4j.discordjson.encoding.possible;
 
+import discord4j.discordjson.possible.Possible;
 import org.immutables.encode.Encoding;
-import reactor.util.annotation.Nullable;
+import javax.annotation.Nullable;
 
 import java.util.Objects;
 import java.util.Optional;
