+++ encoding/src/main/java/discord4j/discordjson/encoding/possible/PossibleOptionalIdEncoding.java
@@ -1,8 +1,9 @@
-package discord4j.discordjson.possible;
+package discord4j.discordjson.encoding.possible;
 
 import discord4j.discordjson.Id;
+import discord4j.discordjson.possible.Possible;
 import org.immutables.encode.Encoding;
-import reactor.util.annotation.Nullable;
+import javax.annotation.Nullable;
 
 import java.util.Objects;
 import java.util.Optional;
